+++ scripts/run_quad_annot.sh
